Auto start, and servers.
Auto Start: The fan now auto starts after 30 seconds after the boot of the jetson. The two camera servers, vision-primary.servie and vision-secondary.service, also auto boot when the jetson turns on.

Server: This is the new camera server that works (please don't mess with it), the server is working with both of the servers and pops up both camera streams.
Co-Authored-By: alex <alex.h@frc1073.org>